BAU: Move EventEmitterConfig into configuration dir
- We have a configuration dir now so move the event emitter
config into there
