+++ hub/policy/src/main/java/uk/gov/ida/hub/policy/PolicyConfiguration.java
@@ -9,6 +9,7 @@ import io.dropwizard.util.Duration;
 import uk.gov.ida.common.ServiceInfoConfiguration;
 import uk.gov.ida.configuration.ServiceNameConfiguration;
 import uk.gov.ida.hub.policy.configuration.SessionStoreConfiguration;
+import uk.gov.ida.hub.policy.configuration.EventEmitterConfiguration;
 import uk.gov.ida.restclient.RestfulClientConfiguration;
 import uk.gov.ida.shared.dropwizard.infinispan.config.InfinispanConfiguration;
 import uk.gov.ida.shared.dropwizard.infinispan.config.InfinispanServiceConfiguration;
